fastapi-application/requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
fastapi
uvicorn
alembic
bcrypt
PyJWT
replicate